requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
